# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Traveloop API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateMe
      tags: [auth]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteMe
      tags: [auth]
      summary: Delete current user account
      responses:
        "204":
          description: No Content
...
  /cities:
    get:
      operationId: listCities
      tags: [cities]
      summary: Search cities
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: country
          schema:
            type: string
        - in: query
          name: region
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/City"
  /cities/{cityId}:
    get:
      operationId: getCity
      tags: [cities]
      summary: Get city by ID
      parameters:
        - in: path
          name: cityId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CityDetail"
...
  /community/{shareCode}:
    get:
      operationId: getPublicTrip
      tags: [community]
      summary: Get a public trip by share code
      parameters:
        - in: path
          name: shareCode
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDetail"
  /community/{shareCode}/copy:
    post:
      operationId: copyPublicTrip
      tags: [community]
      summary: Copy a public trip to current user's trips
      parameters:
        - in: path
          name: shareCode
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [firstName, lastName, email, password]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
        phoneNumber:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, firstName, lastName, email, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        photo:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        language:
          type: string
        isAdmin:
          type: boolean
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        photo:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        language:
          type: string
...
    Trip:
      type: object
      required: [id, userId, name, startDate, endDate, status, isPublic, stopCount, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        description:
          type: ["string", "null"]
        coverPhoto:
          type: ["string", "null"]
        isPublic:
          type: boolean
        shareCode:
          type: ["string", "null"]
        status:
          type: string
          enum: [upcoming, ongoing, completed]
        stopCount:
          type: integer
        totalBudget:
          type: ["number", "null"]
        createdAt:
          type: string
...
    TripDetail:
      type: object
      required: [id, userId, name, startDate, endDate, status, isPublic, stopCount, stops, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        description:
          type: ["string", "null"]
        coverPhoto:
          type: ["string", "null"]
        isPublic:
          type: boolean
        shareCode:
          type: ["string", "null"]
        status:
          type: string
          enum: [upcoming, ongoing, completed]
        stopCount:
          type: integer
        totalBudget:
          type: ["number", "null"]
        stops:
          type: array
          items:
            $ref: "#/components/schemas/Stop"
        createdAt:
          type: string
    Stop:
      type: object
      required: [id, tripId, cityId, cityName, country, startDate, endDate, order]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        cityId:
          type: integer
        cityName:
          type: string
        country:
          type: string
        imageUrl:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        order:
          type: integer
        notes:
          type: ["string", "null"]
        activities:
          type: array
          items:
            $ref: "#/components/schemas/StopActivity"
...
    City:
      type: object
      required: [id, name, country, region, costIndex, popularity]
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
        region:
          type: string
        costIndex:
          type: number
        popularity:
          type: integer
        imageUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
    CityDetail:
      type: object
      required: [id, name, country, region, costIndex, popularity]
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
        region:
          type: string
        costIndex:
          type: number
        popularity:
          type: integer
        imageUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        activities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
    Activity:
      type: object
      required: [id, cityId, cityName, name, type, cost, duration]
      properties:
        id:
          type: integer
        cityId:
          type: integer
        cityName:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
        cost:
          type: number
        duration:
          type: number
        imageUrl:
          type: ["string", "null"]
    StopActivity:
      type: object
      required: [id, stopId, activityId, name, cost]
      properties:
        id:
          type: integer
        stopId:
          type: integer
        activityId:
          type: integer
        name:
          type: string
        type:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
        time:
          type: ["string", "null"]
        cost:
          type: number
        duration:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
...