# backend/src/schemas/generated/types/activity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Traveloop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Activity {
  id: number;
  cityId: number;
  cityName: string;
  name: string;
  /** @nullable */
  description?: string | null;
  type: string;
  cost: number;
  duration: number;
  /** @nullable */
  imageUrl?: string | null;
}