# backend/src/schemas/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Traveloop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  photo?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  country?: string | null;
  language?: string;
  isAdmin?: boolean;
  createdAt: string;
}